[Issue #129] Implementing Realistic Data for All Test Data Generated Entities

+++ backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/basetest/TestData.java
@@ -60,7 +60,7 @@ public interface TestData {
     LocalDateTime END =
         LocalDateTime.of(2020, 11, 13, 18, 15, 0, 0);
     Duration DURATION = Duration.between(START, END);
-    List<Integer> PRICES = List.of(30, 40);
+    List<Double> PRICES = List.of(30.0, 40.0);
     List<Show> SHOWS = new ArrayList<>() {
         {
             add(Show.builder()
